Updates the code with proper alignment
